Sanity checking for directories.
Now there's code to ensure that all the directories in the config files
are actually directories and that they are readable and writeable as
need be, throwing affatal if not.

And, as an added bonus, sitedir and workdir are checked to make sure
they internally end in a forward slash for easier directory name
concatenation.