Removed dependency on Switch.pm.
Looking through the code, we've got exactly one use of Switch.pm.  That
looks like it's a weird case I don't think gets used very often, but the
end result is that it's exactly one use, and it seems Switch.pm isn't
included by default in some Perl distributions (i.e. Ubuntu), making it
one more step in getting it set up.  So, that use has now been unrolled
into an if/elsif/else block and we no longer depend on Switch.pm.

If it turns out to be more convenient later, we can re-add it as needed.