Added footnotes field to JSON output.
The footnotes field contains all the text or HTML stuff that might
wind up in a given day's comics.  This also means days that only
have footnotes won't be completely useless.

Note that this largely assumes footnotes always come after comic
images.  It's technically possible to have a layout that goes from
an image to a footnote to another image, and the footnotes field
does not preserve that situation at the moment.