Added the ***og_images*** tag.
This will add in og:image meta tags for the current comic's images (not
any footnotes).  Any other OpenGraph tags should be added by the user
directly, as those should be static or otherwise can be assembled using
existing template tags.

Okay, granted, having the ability to put in height, width, and type
would be nice, but we don't have that data at the moment.  This is more
of a quick hack anyway.