Added captions to the JSON output.
Now that I know how this code that's been in the script for decades
actually works, there's no longer a "caption" key in the JSON output
that leads to potentially a single string per comic date.  Now it's a
"captions" key that leads to an object with potentially one key per
image file in a comic.  Much better!

Also, the captions are now whitespace trimmed.