Moved caption and alt JSON data into metadata.
Rather than make two object-value pairs for captions and alts, this
introduces a new imgsmetadata field.  This is an array containing the
URLs of each image in the same order as imgs and its corresponding
caption and alt, if they exist.  This way, if a user wants this
metadata, they can just read the array in order and grab everything
without having to cross-reference two more objects.